refactor: Separating commands into files and creating setup function
