+++ bot/commands/endCommand.py
@@ -1,0 +1,13 @@
+from discord.ext import commands
+
+@commands.command()
+async def end(ctx):
+  '''
+    Finaliza o jogo e remove o canal de texto criado
+  '''
+
+  game = ctx.bot.game
+  
+  await game.channel.send('Jogo finalizado!')
+  await game.channel.delete()
+  game.end_game()
