+++ bot/commands/main.py
@@ -1,0 +1,6 @@
+from commands.playCommand import play
+from commands.acceptCommand import accept
+from commands.placeCommand import place
+from commands.endCommand import end
+
+_all__ = (play, accept, place, end)
