+++ bot/commands/playCommand.py
@@ -1,6 +1,5 @@
 from discord.ext import commands
-from context import gameContext
-from models.GameModel import Game
+from models.main import Game
 import utils
 
 @commands.command()
@@ -15,11 +14,21 @@ async def play(ctx, invited_user):
     return
   
   game = Game(challenged_user_id)
-  gameContext.set(game)
 
   if game.state != 0:
     await ctx.send('Há um jogo em andamento, utilize o comando !end para finalizar o jogo')
     return
 
   game.add_user(ctx.author.id)
+  
+  ctx.bot.game = game
   await ctx.send(f'{invited_user} o usuário <@{ctx.author.id}> está te desafiando para uma partida de Jogo da Velha!\nPara aceitar use o comando !accept')
+
+@play.error
+async def handle_play_error(ctx, error):
+  print(error)
+  if isinstance(error, commands.errors.MissingRequiredArgument):
+    await ctx.send('É necessário passar um usuário para jogar com você!')
+  elif isinstance(error, commands.errors.CommandInvokeError):
+    await ctx.send('Usuário convidado é inválido!')
+
