--- bot/context.py
@@ -1,3 +1,0 @@
-import contextvars
-
-gameContext = contextvars.ContextVar('game', default=None)
