+++ bot/main.py
@@ -1,21 +1,15 @@
-from pretty_help import PrettyHelp
 from discord.ext import commands
-# import commands as bot_commands
-from commands import playCommand, acceptCommand
-from dotenv import load_dotenv
+from commands.main import play, accept, place, end
+import settings
 import os
 
-ENDING_NOTE = '''Digite #help command para mais informações sobre o comando.'''
+bot = commands.Bot(command_prefix='')
 
-load_dotenv()
+settings.setup(bot)
 
-COMMAND_PREFIX = '!' if os.getenv('ENVIROMENT') == 'production' else '#'
-
-bot = commands.Bot(command_prefix=COMMAND_PREFIX, help_command=PrettyHelp(ending_note=ENDING_NOTE, show_index=False, no_category='Geral'))
-
-bot.add_command(playCommand.play)
-bot.add_command(acceptCommand.accept)
-# bot.add_command(bot_commands.end)
-# bot.add_command(bot_commands.place)
+bot.add_command(play)
+bot.add_command(accept)
+bot.add_command(place)
+bot.add_command(end)
 
 bot.run(os.getenv('DISCORD_TOKEN'))
