+++ bot/models/main.py
@@ -1,0 +1,3 @@
+from models.GameModel import Game
+
+__all_ = (Game,)
